+++ database/migrations/2022_04_14_013920_create_politic_group_offices_table.php
@@ -17,10 +17,10 @@ class CreatePoliticGroupOfficesTable extends Migration
             $table->id();
             $table->integer('politic_group_id');
             $table->text('address');
-            $table->integer('state_id')->nullable();
-            $table->integer('delegation_id')->nullable();
-            $table->integer('location_id')->nullable();
-            $table->integer('district_id')->nullable();
+            $table->bigInteger('state_id')->nullable();
+            $table->bigInteger('delegation_id')->nullable();
+            $table->bigInteger('location_id')->nullable();
+            $table->bigInteger('district_id')->nullable();
             $table->string('phones')->nullable();
             $table->string('email')->nullable();
             $table->string('schedules')->nullable();
